Updated the linters to be up to date with our current code. Working on a AI copilot.
