+++ Frontend/package.json
@@ -39,11 +39,13 @@
 		"vite": "^5.0.3"
 	},
 	"dependencies": {
+		"@codemirror/autocomplete": "^6.18.6",
 		"@codemirror/basic-setup": "^0.20.0",
-		"@codemirror/language": "^6.10.8",
+		"@codemirror/commands": "^6.8.1",
+		"@codemirror/language": "^6.11.0",
 		"@codemirror/state": "^6.5.2",
 		"@codemirror/theme-one-dark": "^6.1.2",
-		"@codemirror/view": "^6.36.3",
+		"@codemirror/view": "^6.36.5",
 		"@lezer/highlight": "^1.0.0",
 		"@lezer/lr": "^1.0.0",
 		"@sveltejs/adapter-vercel": "^5.6.1",
@@ -51,6 +53,8 @@
 		"bird-site-frontend": "file:",
 		"code-prettify": "^0.1.0",
 		"codemirror": "^6.0.1",
+		"codemirror-copilot": "^0.0.7",
+		"cohere-ai": "^7.16.0",
 		"dotenv": "^16.4.7",
 		"moo": "^0.5.2",
 		"svelte-codemirror-editor": "^1.4.1",
