+++ Frontend/src/lib/codemirror/include/callable.ts
@@ -1,0 +1,1 @@
+export class SemanticCallable {};
