+++ Frontend/src/lib/codemirror/include/sym_table.ts
