+++ Frontend/src/lib/codemirror/include/type.ts
@@ -1,0 +1,1 @@
+export class SemanticType {};
