+++ Frontend/src/lib/codemirror/parser/parser.js
@@ -1,20 +1,20 @@
 // This file was generated by lezer-generator. You probably shouldn't edit it.
 import {LRParser, LocalTokenGroup} from "@lezer/lr"
-import {taggedSyntax} from "../syntax-highlighter"
-const spec_IDENTIFIER = {__proto__:null,fn:14, int:26, float:28, bool:30, str:32, void:34, var:56, true:72, false:74, not:98, as:104, match:128, else:138, and:170, xor:172, or:174, if:182, const:186, print:190, while:194, for:198, return:202, break:206, continue:210, type:216, struct:220}
+import {taggedSyntax} from "../syntax_highlighter"
+const spec_IDENTIFIER = {__proto__:null,fn:14, int:22, float:24, bool:26, str:28, void:30, var:50, self:60, true:66, false:68, and:124, xor:126, or:128, not:134, as:138, match:156, else:160, if:168, const:172, print:176, while:180, for:184, return:188, break:192, continue:196, type:202, struct:206}
 export const parser = LRParser.deserialize({
   version: 14,
-  states: "=|O!yQPOOP#QOPOOO#YO`O'#DSO#eQPO'#C|OOQO'#C|'#C|O%qQPO'#CzO)XQPO'#EjOOQO'#D^'#D^O)`QPO'#D]O%qQPO'#CzO*TQPO'#DlOOQO'#Dt'#DtO)`QPO'#EWOOQO'#Cz'#CzOOQO'#Cv'#CvO*[QPO'#CtOOQO'#C`'#C`O+UQPO'#C`OOQO'#FO'#FOO+ZQPO'#E}QOQPOOOOQO'#DP'#DPO)`QPO'#DmO+bQPO'#CwO+gQPO'#EYO,[QPO'#E[O)`QPO'#E^O+gQPO'#E`O,aQPO'#EbO-ZQPO'#EdO-bQPO'#EfO-gQPO'#EhO-lQPO'#EkO-qQPO'#CaO-vQPO'#EmP-{OQO'#C]P.TOWO'#C^POOO)C@j)C@jOOOO'#Er'#ErO.`O`O,59nOOQO,59n,59nO.kQPO,5:SOOQO'#DT'#DTO)`QPO,59gO)`QPO,59uO.sQPO,59}O)`QPO,5:OO.zQPO,5:QOOQO'#Du'#DuO)`QPO,5;uOOQO'#Dx'#DxO)`QPO,5;vOOQO'#D}'#D}O)`QPO,5;wOOQO'#EP'#EPO)`QPO,5;xOOQO,5;U,5;UO/PQPO,59{O)`QPO,5;yO)`QPO,5;zO)`QPO,5;{OOQO,59w,59wO)`QPO,5:PO/eQPO'#FYO/uQPO'#FXO0QQPO,5:WO0VQPO,5:rOOQO'#FS'#FSO0^QPO'#FRO1WQPO,59`OOQO,58z,58zOOQO,5;j,5;jO2gQPO,5:XO2nQPO,59cO2vQPO,5:tO4WQPO,5:vO4`QPO,5:xO2vQPO,5:zOOQO'#Fb'#FbO4hQPO,5:|O4oQPO,5;OOOQO,5;O,5;OOOQO,5;Q,5;QOOQO,5;S,5;SO4vQPO,5;VO4{QPO,58{O5QQPO,5;XPOOO'#Ep'#EpP5VOQO,58wPOOO,58w,58wPOOO'#Eq'#EqP5_OWO,58xPOOO,58x,58xOOOO-E8p-E8pOOQO1G/Y1G/YO5jQPO'#DjO5oQPO'#DiO5wQPO1G/nOOQO1G/R1G/RO5|QPO1G/aO6TQPO1G/iO6YQPO1G/jOOQO1G/l1G/lOOQO1G1a1G1aOOQO1G1b1G1bOOQO1G1c1G1cOOQO1G1d1G1dOOQO'#Cg'#CgOOQO1G/g1G/gOOQO'#Ch'#ChOOQO1G1e1G1eOOQO1G1f1G1fOOQO1G1g1G1gOOQO1G/k1G/kO)`QPO,5;tOOQO1G/r1G/rOOQO1G0^1G0^OOQO,5;n,5;nOOQO1G.z1G.zO6aQPO1G/sO)`QPO1G.}O/PQPO1G.}O6hQPO1G0`O)`QPO1G0bO/PQPO1G0bOOQO1G0d1G0dOOQO1G0f1G0fO8bQPO'#FcO8iQPO1G0hOOQO1G0j1G0jO/PQPO1G0qO8nQPO1G.gO8vQPO1G0sPOOO-E8n-E8nPOOO1G.c1G.cPOOO-E8o-E8oPOOO1G.d1G.dO)`QPO,5:UO9OQPO,5:TOOQO7+%Y7+%YO)`QPO7+${OOQO7+%T7+%TOOQO7+%U7+%UO9WQPO,59RO9]QPO1G1`O9mQPO1G/jO9wQPO'#DoO:PQPO7+%_O:UQPO7+$iO:]QPO7+$iO:eQPO7+%zO:mQPO7+%|O:tQPO7+%|O:|QPO7+&SO;TQPO7+&]O;]QPO'#CeOOQO'#FQ'#FQO;bQPO'#FPO;jQPO7+$RO;oQPO'#EoOOQO'#Fe'#FeO;tQPO'#FdO;|QPO7+&_O<RQPO1G/pO<]QPO1G/pOOQO<<Hg<<HgOOQO1G.m1G.mO)`QPO,5:[O<bQPO,5:ZO<iQPO'#DrO<nQPO<<HyOOQO<<HT<<HTO)`QPO<<HTOOQO<<If<<IfOOQO<<Ih<<IhO)`QPO<<IhO<sQPO<<InOOQO<<Iw<<IwO/PQPO,59PO<xQPO,5;lO<}QSO<<GmO/PQPO,5;ZO=VQPO,5<OOOQO<<Iy<<IyO)`QPO7+%[O=_QPO1G/vO=iQPO1G/vO)`QPO,5:^OOQOAN>eAN>eO=pQPOAN=oO=wQPOAN?SOOQOAN?YAN?YO>OQPO1G.kOOQO1G1W1G1WO/PQPO'#CrO<sQPOAN=XO>ZQPO1G0uOOQO1G1k1G1kO>fQPO<<HvO)`QPO7+%bO>pQPO1G/xOOQOG23ZG23ZOOQOG24nG24nO>wQPO,59^OOQOG22sG22sO?PQPO<<H|O+gQPO'#D]O+gQPO,59gO+gQPO,5;uO+gQPO,5;vO+gQPO,5;wO+gQPO,5;xO+gQPO,5;yO+gQPO,5;zO+gQPO,5;{O+gQPO,5:PO@[QPO'#FcO+gQPO7+${O@oQPO'#CzO@oQPO'#CzOBcQPO1G/aO)`QPO,59u",
-  stateData: "Bm~O#pOS#iPQ#lPQ~OUROVqOW[ObYOi_OlgOqSOrSOteOueO!RVO!SVO!bfO!}hO#PiO#RjO#TkO#VlO#XmO#ZnO#]oO#`pO#brO#wQO~O#g#qP~P]O#isO#ltO~O#xvO#yvO#zxO~OiyOWpXbpXmpXxpXypXzpX{pX|pX!OpX!SpX!UpX!ZpX!jpX!kpX!mpX!npX!opX!ppX!rpX!tpX!upX!vpX!wpX!xpX!ypX!{pXcpXdpXepXYpX!epX!_pX!gpX~OmzOxzOyzOzzO{zO|zOWnXbnX!OnX!SnX!UnX!ZnX!jnX!knX!mnX!nnX!onX!pnX!rnX!tnX!unX!vnX!wnX!xnX!ynX!{nXcnXdnXenXYnX!enX!_nX!gnX~OW}Ob!OO!O|O!S!UO!U!ZO!Z!PO!j!QO!k!QO!m!SO!n!SO!o!SO!p!SO!r!UO!t!WO!u!WO!v!WO!w![O!x!]O!y!^O~O!{!YO~P'zOUROW[ObYOqSOrSOteOueO!RVO!SVO!bfO#wQO~Oc#{P~P)`Oi_OlgO!}hO#PiO#RjO#TkO#VlO#XmO#ZnO#]oO#`pO!_#uP~P)`O!{!hO~O#g#qX~P]OU!kO~OUSOW[ObYOqSOrSOteOueO!RVO!SVO!bfO#wQO~OU!mO~Oi_OlgO!{!pO!}hO#PiO#RjO#TkO#VlO#XmO#ZnO#]oO#`pO~P)`O!{!sO~P)`O!{!tO~O!{!uO~OU!vO~OU!wO~OU!xO~O#j!yO#k!{O~O#m!|O#n!|O#o#OO~O#xvO#yvO#z#QO~OU#RO!_!]P~Oe#{P~P)`OU#YO~OU#_O]#aO^#aO_#aO`#aOa#aO~Oc#|Xd#|X!{#|Xe#|X~P'zOd#fOc#{Xe#{X~Oc#gO~Oe#hO~P'zOi_OlgO!}hO#PiO#RjO#TkO#VlO#XmO#ZnO#]oO#`pO!_#uX~P)`O!_#jO~OW}O!O|O!S!UO!U!ZO!Z!PO!j!QO!k!QO!m!SO!n!SO!o!SO!p!SO!r!UO!t!WO!u!WO!v!WO!w![O!x!]O!y!^O~Ob#kO~P1]OY#mOm#lO~OW}Ob!OOi_O!O%uO!S!UO!U!ZO!Z!PO!j!QO!k!QO!m!SO!n!SO!o!SO!p!SO!r!UO!t!WO!u!WO!v!WO!w%lO!x%mO!y%nO~OY#pOm#oO~Od#fO!{#qO~O!{$VP~P)`O!{#uO~P'zOm#vO~OW#wO~Oi#xO~O#j!yO#k#zO~O#m!|O#n!|O#o#|O~Om#}O~Od$OO!_!]X~O!_$PO~OY$QO~P'zOe$RO~Oc$SO~P'zO!g!cP~P)`O!g$[OU!|iV!|iW!|ib!|ii!|il!|iq!|ir!|it!|iu!|i!R!|i!S!|i!b!|i!}!|i#P!|i#R!|i#T!|i#V!|i#X!|i#Z!|i#]!|i#`!|i#b!|i#g!|i#w!|i!_!|i!{!|i~O!{$VX~P'zO!{$_O~OU$aOe#sP~OU$eO!_$WP~OU$jO!_!]a~Oc$lO~Oc#|id#|i!{#|ie#|i~P'zOc$SO!e$mO~P'zOd$nO!g!cX~O!g$oO~O!{$qO~P'zOb$TOm$rO~Oi_O!}hO~O!{$tO~P'zOb$TOm$uO~Oi$VP~P+gOb$TO!{$wO~OY$xO~Od$yOe#sX~Oe$zO~OY${O~Od$|O!_$WX~O!_$}O~Od!^i!_!^i~P'zOm%OO~O!g!ca~P)`O!e%RO~Oc%SO~Oi_O~OU$aO~Og%YOifP~OU$eO!_$Wa~Od!di!g!di~P'zO!e%_O~P'zO!{%aO~P'zO!{%bO~P'zOb$TOdXieXi~Ob$TOd#ci!_#ci~Od!^y!_!^y~P'zOc!fi~P'zOb$TOifa~Od!dy!g!dy~P'zOW}Ob!OO!S!UO!U!ZO!Z!PO!j!QO!k!QO!m!SO!n!SO!o!SO!p!SO!r!UO!t!WO!u!WO!v!WO~O!O%uO!w%lO!x%mO!y%nOi$VX~P?ZOmzOxzOyzOzzO{zO|zOWnXbnXinX!OnX!SnX!UnX!ZnX!jnX!knX!mnX!nnX!onX!pnX!rnX!tnX!unX!vnX!wnX!xnX!ynX~OY%qO~P'zOrq~",
-  goto: "4|$YP$Z$ZP$^$dPPP$hP$n%TPPPPPPPPP%^P%aP%w&UPP&]({*TPP+_PP+_,gPPPPP({P({,sPP({P({-}({({P({/X/[P({({P/_/bP/eP({/hPP0UPPPP0rP1`PPPPPP({P1|P&UP&UP&UP&UP&UP&UP&UP&U&UP2WP2[2b2h2nPPPPPPPPPP2t2w2z2}3Q3TPPPP3W3^3e3e3e3e3e3e3e4m4p4v4yRuPQbOR!icT`OcQ$b#wR%X$yQ#`!ZQ$Z#mQ$^#pQ$`#vQ%W$xQ%[${R%c%Y_#_!Z#m#p#v$x${%YR%Z$zY^O_cl!fQ#n!lQ#r!oQ$s$[Q%V$vR%d%ZS`OcQ!e_Q!plR#i!fZ^O_cl!fYUO_cl!fS!_W%fU!aYj}Q!d[Q!jfQ!lhQ!okQ!rmS#U{%gQ#V|Q#X!OS#Z!R%hS#[!T%iS#]!V%jS#^!X%kS#b![%lS#c!]%mS#d!^%nS#e!`%oQ#s!qQ$U#fQ$V#kQ$Y#lQ$]#oQ$i#}S$k$Q%qQ%P$mQ%Q$nQ%T$rQ%U$uQ%^%OQ%`%RQ%e%_Q%p$_R%t%u#]]OWY[_cfhjklm{|}!O!R!T!V!X![!]!^!`!f!q#f#k#l#o#}$Q$_$m$n$r$u%O%R%_%f%g%h%i%j%k%l%m%n%o%q%u!nTOWY[_cfjlm{|}!O!R!T!V!X![!]!^!`!f!q#f#k#l#o#}$Q$m$n$r$u%O%R%_%um%rhk$_%f%g%h%i%j%k%l%m%n%o%q#]SOWY[_cfhjklm{|}!O!R!T!V!X![!]!^!`!f!q#f#k#l#o#}$Q$_$m$n$r$u%O%R%_%f%g%h%i%j%k%l%m%n%o%q%uQ{TQ!`XQ%g%rR%o%s!nWOWY[_cfjlm{|}!O!R!T!V!X![!]!^!`!f!q#f#k#l#o#}$Q$m$n$r$u%O%R%_%um%fhk$_%f%g%h%i%j%k%l%m%n%o%q!nXOWY[_cfjlm{|}!O!R!T!V!X![!]!^!`!f!q#f#k#l#o#}$Q$m$n$r$u%O%R%_%um%shk$_%f%g%h%i%j%k%l%m%n%o%qR#TyR#SyR$X#kR$W#kR$p$Xz!RU!a!d!j!r#V#X#s$U$V$Y$]$i%P%Q%T%U%^%`%e%tV%h!l!o%pz!TU!a!d!j!r#V#X#s$U$V$Y$]$i%P%Q%T%U%^%`%e%tV%i!l!o%pz!VU!a!d!j!r#V#X#s$U$V$Y$]$i%P%Q%T%U%^%`%e%tV%j!l!o%pz!XU!a!d!j!r#V#X#s$U$V$Y$]$i%P%Q%T%U%^%`%e%tV%k!l!o%pY^O_cl!fR$s$[TaOcQ$f#xR%]$|Q!zsR#y!zQ!}tR#{!}QwQR#PwRdORcOR$d#wR$c#wR!g_R!f_Q!cYR#W}S!bY}R!nj#]ZOWY[_cfhjklm{|}!O!R!T!V!X![!]!^!`!f!q#f#k#l#o#}$Q$_$m$n$r$u%O%R%_%f%g%h%i%j%k%l%m%n%o%q%uR!qlQ#t!qR$v$_R$h#xR$g#x",
-  nodeNames: "⚠ SINGLELINE_COMMENT MULTILINE_COMMENT Program Stmt Func IDENTIFIER fn LPAREN Param COLON Type_identifier TYPE_LITERAL int float bool str void LBRACKET RBRACKET COMMA RPAREN Return_type ARROW Block LBRACE Block_valid_stmt Decl_stmt var EQUAL Expr Assign_expr Primary INT_LITERAL FLOAT_LITERAL BOOL_LITERAL true false STR_LITERAL ASSIGN_OP PLUS_EQUAL MINUS_EQUAL STAR_EQUAL SLASH_EQUAL PERCENT_EQUAL Ternary_expr QUESTION Unary_expr PREFIX_UNARY_OP not MINUS Type_cast as Call_expr Subscript_expr Index_assign Direct_member_access DOT Struct_initialization Maybe_struct_initialization_list Struct_initialization_list RBRACE Array_initialization Match match Maybe_match_arms Match_arms FAT_ARROW Match_else_arm else Binary_expr EQUALITY_OP EQUAL_EQUAL BANG_EQUAL COMPARISON_OP GREATER GREATER_EQUAL LESS LESS_EQUAL TERM_OP PLUS FACTOR_OP STAR SLASH PERCENT and xor or Grouping SEMICOLON If_stmt if Const_stmt const Print_stmt print While_stmt while For_stmt for Return_stmt return Break_stmt break Continue_stmt continue Expr_stmt Type_stmt type Struct_decl struct Field_member",
-  maxTerm: 147,
+  states: "=|O!|QPOOP#TOPOOO#]O`O'#DPO%XQPO'#CyOOQO'#Cy'#CyO'SQPO'#EcO'ZQPO'#CwOOQO'#D['#D[OOQO'#Dp'#DpO)aQPO'#DoO'ZQPO'#CwO*XQPO'#DzO)aQPO'#EPOOQO'#Cw'#CwOOQO'#Cs'#CsO*`QPO'#CqOOQO'#C`'#C`OOQO'#Ew'#EwO+YQPO'#EvQOQPOOO+aQPO'#D{O,XQPO'#CtO+aQPO'#ERO,^QPO'#ETO)aQPO'#EVO+aQPO'#EXO,cQPO'#EZO-]QPO'#E]O-dQPO'#E_O-iQPO'#EaO-nQPO'#EdO-sQPO'#CaO-xQPO'#EfP-}OQO'#C]P.VOWO'#C^POOO)C@c)C@cOOOO'#Ek'#EkO.bO`O,59kOOQO,59k,59kO.mQPO,5:`O.tQPO,5:dO.|QPO,59rO)aQPO,59tOOQO'#D]'#D]O)aQPO,5;nOOQO'#D`'#D`O)aQPO,5;oOOQO'#De'#DeO)aQPO,5;pOOQO'#Dh'#DhO)aQPO,5;qO)aQPO,5:bOOQO,5:},5:}O)aQPO,5;rO)aQPO,5;sO)aQPO,5;tO/RQPO,5:^OOQO'#DQ'#DQO)aQPO,59dOOQO,5:Z,5:ZO)aQPO,5:cO/gQPO'#F[O/wQPO'#FZO0SQPO,5:fO0XQPO,5:kOOQO'#E}'#E}O0`QPO'#E|O1YQPO,59]OOQO,5;c,5;cO2iQPO,5:gO2pQPO,59`O2xQPO,5:mO3PQPO,5:oO3XQPO,5:qO2xQPO,5:sOOQO'#Fb'#FbO3aQPO,5:uO3hQPO,5:wOOQO,5:w,5:wOOQO,5:y,5:yOOQO,5:{,5:{O3oQPO,5;OO3tQPO,58{O3yQPO,5;QPOOO'#Ei'#EiP4OOQO,58wPOOO,58w,58wPOOO'#Ej'#EjP4WOWO,58xPOOO,58x,58xOOOO-E8i-E8iOOQO1G/V1G/VO4cQPO1G/zO4hQPO'#F^O4mQPO'#F]O4uQPO1G0OO4zQPO1G/^OOQO1G/{1G/{O7WQPO1G/`OOQO1G1Y1G1YOOQO1G1Z1G1ZOOQO1G1[1G1[OOQO1G1]1G1]O7_QPO1G/|OOQO1G1^1G1^OOQO1G1_1G1_OOQO1G1`1G1`OOQO'#Cf'#CfOOQO1G/x1G/xOOQO1G/O1G/OOOQO1G/}1G/}O)aQPO,5;vOOQO1G0Q1G0QOOQO1G0V1G0VOOQO,5;i,5;iOOQO1G.w1G.wO7fQPO1G0RO)aQPO1G.zO/RQPO1G.zO7mQPO1G0XO)aQPO1G0ZO/RQPO1G0ZOOQO1G0]1G0]OOQO1G0_1G0_O9jQPO'#FcO9qQPO1G0aOOQO1G0c1G0cO/RQPO1G0jO9vQPO1G.gO:OQPO1G0lPOOO-E8g-E8gPOOO1G.c1G.cPOOO-E8h-E8hPOOO1G.d1G.dOOQO7+%f7+%fO)aQPO,5;xO:ZQPO,5;wOOQO7+%j7+%jO)aQPO7+$zOOQO7+%h7+%hO:cQPO,59QO:hQPO1G1bO:xQPO'#F`O;PQPO'#F_O;XQPO7+%mO;^QPO7+$fO;eQPO7+$fO;mQPO7+%sO;uQPO7+%uO;|QPO7+%uO<UQPO7+%{O<]QPO7+&UO<eQPO'#EzOOQO'#Ey'#EyO<jQPO'#ExO<rQPO7+$RO<wQPO'#EhOOQO'#Ff'#FfOOQO'#Fe'#FeO<|QPO'#FdO=XQPO7+&WO=^QPO1G1dO=hQPO1G1dOOQO<<Hf<<HfOOQO1G.l1G.lO)aQPO,5;zO=mQPO,5;yO=tQPO'#FaO=yQPO<<IXOOQO<<HQ<<HQO)aQPO<<HQOOQO<<I_<<I_OOQO<<Ia<<IaO)aQPO<<IaO>OQPO<<IgOOQO<<Ip<<IpO/RQPO,5;fO>TQPO,5;eO>YQSO<<GmO/RQPO,5;SOOQO,5<P,5<PO>bQPO<<IrO)aQPO7+'OO>gQPO1G1fO>qQPO1G1fO)aQPO,5;{OOQOAN>sAN>sO>xQPOAN=lO?PQPOAN>{OOQOAN?RAN?RO?WQPO1G1QOOQO1G1P1G1PO/RQPO'#E{O>OQPOAN=XO?cQPO1G0nOOQOAN?^AN?^O?kQPO<<JjO)aQPO7+'QO?uQPO1G1gOOQOG23WG23WOOQOG24gG24gO?|QPO,5;gOOQOG22sG22sOOQO7+&Y7+&YO@UQPO<<JlO@`QPO'#CyO+aQPO'#DoO+aQPO,5;nO+aQPO,5;oO+aQPO,5;pO+aQPO,5;qO+aQPO,5;rO+aQPO,5;sO+aQPO,5;tO+aQPO,59dO+aQPO,5:cO@gQPO'#FcO+aQPO7+$zO@nQPO'#CwO@nQPO'#CwOB_QPO1G/`O)aQPO,59t",
+  stateData: "Bi~O#iOS#bPQ#ePQ~OUROVoOW[O`ZOf_OieOnSOoSOpSOqSOrSO!ZWO!eWO!pdO!vfO!xgO!zhO!|iO#OjO#QkO#SlO#UmO#XnO#ZpO#rQO~O#`#jP~P]O#bqO#erO~O#stO#ttO#uvO~OWwO`mXjmXumXvmXwmXxmXymX{mX}mX!QmX!RmX!TmX!UmX!VmX!WmX!YmX!ZmX!]mX!^mX!_mX!`mX!amX!bmX!gmX~OfxO!tmXamXbmXcmXXmX!qmX!mmX!rmX~P#hO`!TO{yO}zO!Q{O!R{O!T}O!U}O!V}O!W}O!Y!PO!Z!PO!]!RO!^!RO!_!RO!`!VO!a!WO!b!XO!g!YO~O!t!UO~P%xOj!ZOu!ZOv!ZOw!ZOx!ZOy!ZO`kX{kX}kX!QkX!RkX!TkX!UkX!VkX!WkX!YkX!ZkX!]kX!^kX!_kX!`kX!akX!bkX!gkX!tkXakXbkXckXXkX!qkX!mkX!rkX~OUROW[O`ZOnSOoSOpSOqSOrSO!ZWO!eWO!pdO#rQO~Oa#}P~P)aOf_OieO!vfO!xgO!zhO!|iO#OjO#QkO#SlO#UmO#XnO!m#pP~P)aO#`#jX~P]OU%eOW[O`ZOnSOoSOpSOqSOrSO!ZWO!eWO!pdO#rQO~OU!hO~OU!jO~Of_OieO!t!mO!vfO!xgO!zhO!|iO#OjO#QkO#SlO#UmO#XnO~P)aO!t!pO~P)aO!t!qO~O!t!rO~OU!sO~OU!tO~OU!uO~O#c!vO#d!xO~O#f!yO#g!yO#h!{O~O#stO#ttO#u!}O~Oc#}P~P)aOU#PO!m$PP~OU#SO~OU#_OZ#_O[#_O]#_O^#_O_#_O~Oa$OXb$OX!t$OXc$OX~P%xOb#cOa#}Xc#}X~Oa#dO~Oc#eO~P%xOf_OieO!vfO!xgO!zhO!|iO#OjO#QkO#SlO#UmO#XnO!m#pX~P)aO!m#gO~O`!TO{yO}%uO!Q{O!R{O!T}O!U}O!V}O!W}O!Y!PO!Z!PO!]!RO!^!RO!_!RO!`%kO!a%lO!b%mO!g!YO~Of#hO~P1_OX#jOj#iO~Of_O~P1_OX#mOj#lO~Ob#cO!t#nO~O!t$VP~P)aO!t#rO~P%xOj#sO~OW#tO~Of#uO~O#c!vO#d#wO~O#f!yO#g!yO#h#yO~Oc#zO~Oj#{O~Ob#|O!m$PX~O!m#}O~OWwO`zijziuzivziwzixziyzi{zi}zi!Qzi!Rzi!Tzi!Uzi!Vzi!Wzi!Yzi!Zzi!]zi!^zi!_zi!`zi!azi!bzi!gzi!tziazibziczifziXzi!qzi!mzi!rzi~OX$OO~P%xOa$PO~P%xO!r$RP~P)aO!r$XOU!uiV!uiW!ui`!uif!uii!uin!uio!uip!uiq!uir!ui!Z!ui!e!ui!p!ui!v!ui!x!ui!z!ui!|!ui#O!ui#Q!ui#S!ui#U!ui#X!ui#Z!ui#`!ui#r!ui!m!ui!t!ui~O!t$VX~P%xO!t$[O~OU$^Oc#lP~OU$bOVoO!m$WP~OU$hO!m$Pa~Oa$jO~Oa$Oib$Oi!t$Oic$Oi~P%xO!q$kO~P%xOb$lO!r$RX~O!r$mO~O!t$oO~P%xO`$QOj$pO~Of_O!vfO~O!t$rO~P%xO`$QOj$sO~Of$VP~P+aO`$QO!t$uO~OX$vO~Ob$wOc#lX~Oc$xO~OX$yO~OU$bOVoO!m$WX~O!m${O~Ob$Qi!m$Qi~P%xOj$|O~O!r$Ra~P)aO!q%PO~O!m%QO~Of_O~OU$^O~Od%WOf#oP~O!t%ZO~Ob$Si!r$Si~P%xO!q%]O~P%xO!t%_O~P%xO!t%`O~P%xO`$QOb#nic#ni~O`$QO!t%cO~Ob$Qy!m$Qy~P%xO!m$Ti~P%xO`$QOf#oa~Ob$Sy!r$Sy~P%xOfmX~P#hOf$VX~P1_Oj!ZOu!ZOv!ZOw!ZOx!ZOy!ZO`kXfkX{kX}kX!QkX!RkX!TkX!UkX!VkX!WkX!YkX!ZkX!]kX!^kX!_kX!`kX!akX!bkX!gkX~OX%qO~P%xOpo~",
+  goto: "6T$ZP$[$[P$_$ePPPP$mPPPPPPPPPP%SP%j%wPP&O(n)vPPPPP+Q,YPPPPP)vP(nP(n,fPP-SPPPP-pPP.^PPPPPP(n.zP(nP0U(n1a(n(nP(n(nPPP(nP2kP%wP%wP%wP%wP%wP%wP%wP%w%wP2uP$i2y3P3VPPPPPPPPPP3]3`3c3f3i3o3r3uPPPP3x3x3x3x3x3x3x5Q5W5_5b5e5h5k5n5q5w5z5}RsPQaOR!fbS`ObT$c#u$eQ#`!YQ$W#jQ$Z#mQ$]#sQ%U$vQ%Y$yR%a%WY^O_bj!dQ#k!iQ#o!lQ$q$XQ%T$tR%b%XS`ObQ!c_Q!mjR#f!dZ^O_bj!dYTO_bj!dS!]X%fU!_ZhwQ!b[Q!gdQ!ifQ!liQ!okQ#UzS#V|%gS#W!O%hS#X!Q%iS#Y!S%jQ#Z!TS#[!V%kS#]!W%lS#^!X%mS#a![%nS#b!^%oQ#p!nQ$R#cQ$S#hQ$V#iQ$Y#lQ$g#{S$i$O%qQ$}$kQ%O$lQ%R$pQ%S$sQ%[$|Q%^%PQ%d%]Q%p$[R%t%u#]]OXZ[_bdfhijkwz|!O!Q!S!T!V!W!X![!^!d!n#c#h#i#l#{$O$[$k$l$p$s$|%P%]%f%g%h%i%j%k%l%m%n%o%q%u!lUOXZ[_bhjkwz|!O!Q!S!T!V!W!X![!^!d!n#c#h#i#l#{$O$k$l$p$s$|%P%]%uo%rdfi$[%f%g%h%i%j%k%l%m%n%o%q#]SOXZ[_bdfhijkwz|!O!Q!S!T!V!W!X![!^!d!n#c#h#i#l#{$O$[$k$l$p$s$|%P%]%f%g%h%i%j%k%l%m%n%o%q%uQ![UQ!^YQ%n%rR%o%sx|T!_!b!o#U#Z#p$R$S$V$Y$g$}%O%R%S%[%^%d%tX%g!g!i!l%px!OT!_!b!o#U#Z#p$R$S$V$Y$g$}%O%R%S%[%^%d%tX%h!g!i!l%px!QT!_!b!o#U#Z#p$R$S$V$Y$g$}%O%R%S%[%^%d%tX%i!g!i!l%px!ST!_!b!o#U#Z#p$R$S$V$Y$g$}%O%R%S%[%^%d%tX%j!g!i!l%p!lXOXZ[_bhjkwz|!O!Q!S!T!V!W!X![!^!d!n#c#h#i#l#{$O$k$l$p$s$|%P%]%uo%fdfi$[%f%g%h%i%j%k%l%m%n%o%q#[]OXZ[_bdfhijkwz|!O!Q!S!T!V!W!X![!^!d!n#c#h#i#l#{$O$[$k$l$p$s$|%P%]%f%g%h%i%j%k%l%m%n%o%q%uR#Ty!lYOXZ[_bhjkwz|!O!Q!S!T!V!W!X![!^!d!n#c#h#i#l#{$O$k$l$p$s$|%P%]%uo%sdfi$[%f%g%h%i%j%k%l%m%n%o%qY^O_bj!dR$q$XT`ObQ!wqR#v!wQ!zrR#x!zQuQR!|uRcORbOR$a#tR$`#tQ$_#tR%V$wR%X$xR!e_R!d_#]VOXZ[_bdfhijkwz|!O!Q!S!T!V!W!X![!^!d!n#c#h#i#l#{$O$[$k$l$p$s$|%P%]%f%g%h%i%j%k%l%m%n%o%q%uQ!aZR#OwS!`ZwR!khR#RxR#QxR$U#hR$T#hR$n$UR!njQ#q!nR$t$[R$f#uR$e#uQ$d#uR$z$e",
+  nodeNames: "⚠ SINGLELINE_COMMENT MULTILINE_COMMENT Program Top_level_stmt Func IDENTIFIER fn LPAREN COLON Type_identifier int float bool str void LBRACKET RBRACKET COMMA RPAREN ARROW Block LBRACE Stmt Decl_stmt var EQUAL Expr Assign_expr Primary self INT_LITERAL FLOAT_LITERAL true false STR_LITERAL ASSIGN_OP PLUS_EQUAL MINUS_EQUAL STAR_EQUAL SLASH_EQUAL PERCENT_EQUAL Direct_member_access DOT Ternary_expr QUESTION Binary_expr EQUALITY_OP EQUAL_EQUAL BANG_EQUAL COMPARISON_OP GREATER GREATER_EQUAL LESS LESS_EQUAL TERM_OP PLUS MINUS FACTOR_OP STAR SLASH PERCENT and xor or Unary_expr PREFIX_UNARY_OP not Type_cast as Call_expr Method_call Subscript_expr Index_assign Struct_initialization RBRACE Array_initialization Match match FAT_ARROW else Grouping SEMICOLON If_stmt if Const_stmt const Print_stmt print While_stmt while For_stmt for Return_stmt return Break_stmt break Continue_stmt continue Expr_stmt Type_stmt type Struct_decl struct Prop_decl",
+  maxTerm: 148,
   propSources: [taggedSyntax],
-  skippedNodes: [0,1,2,112,113],
+  skippedNodes: [0,1,2,105,106],
   repeatNodeCount: 3,
-  tokenData: ":WlRzOX#uXY$^YZ%fZ]#u]^$^^p#upq$^qr%yrs&usu#uuv'Yvx#uxy(Wyz(kz{)O{|)||}*z}!O+_!O!P,s!P!Q-W!Q![/Y![!]3O!]!^3c!^!_3v!_!`4t!`!a6Y!a!b7W!b!c#u!c!}7k!}#O8h#O#P#u#P#Q8{#Q#R#u#R#S7k#S#T#u#T#o7k#o#p9`#p#q#u#q#r9s#r;'S#u;'S;=`$W<%lO#uW#zS#jWOY#uZ;'S#u;'S;=`$W<%lO#uW$ZP;=`<%l#ul$eY#jW#pdOX#uXY$^YZ%TZ]#u]^$^^p#upq$^q;'S#u;'S;=`$W<%lO#ud%YS#pdXY%TYZ%T]^%Tpq%Tl%mS#kW#pdXY%TYZ%T]^%Tpq%T[&OU#jWOY#uZ!_#u!_!`&b!`;'S#u;'S;=`$W<%lO#u[&iS!kS#jWOY#uZ;'S#u;'S;=`$W<%lO#u[&|S#wS#jWOY#uZ;'S#u;'S;=`$W<%lO#u['aU!vS#jWOY#uZ!_#u!_!`'s!`;'S#u;'S;=`$W<%lO#u['zS|S#jWOY#uZ;'S#u;'S;=`$W<%lO#u[(_SWS#jWOY#uZ;'S#u;'S;=`$W<%lO#u[(rSeS#jWOY#uZ;'S#u;'S;=`$W<%lO#u[)VU!tS#jWOY#uZ!_#u!_!`)i!`;'S#u;'S;=`$W<%lO#u[)pSzS#jWOY#uZ;'S#u;'S;=`$W<%lO#u[*TU!rS#jWOY#uZ!_#u!_!`*g!`;'S#u;'S;=`$W<%lO#u[*nSxS#jWOY#uZ;'S#u;'S;=`$W<%lO#u[+RSdS#jWOY#uZ;'S#u;'S;=`$W<%lO#ul+fV!SS#jWOY#uZ!_#u!_!`+{!`!a,`!a;'S#u;'S;=`$W<%lO#u[,SSyS#jWOY#uZ;'S#u;'S;=`$W<%lO#uh,gSg`#jWOY#uZ;'S#u;'S;=`$W<%lO#u[,zS!ZS#jWOY#uZ;'S#u;'S;=`$W<%lO#ul-_Y!uS#jWOY#uZz#uz{-}{!P#u!P!Q.b!Q!_#u!_!`.u!`;'S#u;'S;=`$W<%lO#ul.US#ld#jWOY#uZ;'S#u;'S;=`$W<%lO#ul.iS#id#jWOY#uZ;'S#u;'S;=`$W<%lO#u[.|S{S#jWOY#uZ;'S#u;'S;=`$W<%lO#u[/aW#jWqSOY#uZ!O#u!O!P/y!P!Q#u!Q![/Y![;'S#u;'S;=`$W<%lO#u[0OU#jWOY#uZ!Q#u!Q![0b![;'S#u;'S;=`$W<%lO#u[0iY#jWrSOY#uZ!Q#u!Q![0b![!g#u!g!h1X!h#X#u#X#Y1X#Y;'S#u;'S;=`$W<%lO#u[1^Y#jWOY#uZ{#u{|1||}#u}!O1|!O!Q#u!Q![2e![;'S#u;'S;=`$W<%lO#u[2RU#jWOY#uZ!Q#u!Q![2e![;'S#u;'S;=`$W<%lO#u[2lU#jWrSOY#uZ!Q#u!Q![2e![;'S#u;'S;=`$W<%lO#u[3VSYS#jWOY#uZ;'S#u;'S;=`$W<%lO#u[3jS!{S#jWOY#uZ;'S#u;'S;=`$W<%lO#u[3}U!oS#jWOY#uZ!_#u!_!`4a!`;'S#u;'S;=`$W<%lO#u[4hS!pS#jWOY#uZ;'S#u;'S;=`$W<%lO#u[4{VmS#jWOY#uZ!_#u!_!`5b!`!a5u!a;'S#u;'S;=`$W<%lO#u[5iS!jS#jWOY#uZ;'S#u;'S;=`$W<%lO#u[5|S!eS#jWOY#uZ;'S#u;'S;=`$W<%lO#u[6aU!mS#jWOY#uZ!_#u!_!`6s!`;'S#u;'S;=`$W<%lO#u[6zS!nS#jWOY#uZ;'S#u;'S;=`$W<%lO#u[7_S!OS#jWOY#uZ;'S#u;'S;=`$W<%lO#u[7r[US#jWOY#uZ!Q#u!Q![7k![!c#u!c!}7k!}#R#u#R#S7k#S#T#u#T#o7k#o;'S#u;'S;=`$W<%lO#u[8oSbS#jWOY#uZ;'S#u;'S;=`$W<%lO#u[9SScS#jWOY#uZ;'S#u;'S;=`$W<%lO#ul9gSid#jWOY#uZ;'S#u;'S;=`$W<%lO#u[9zS!_S#jWOY#uZ;'S#u;'S;=`$W<%lO#u",
-  tokenizers: [2, 3, 4, new LocalTokenGroup("j~RQYZXz{^~^O#n~~aP!P!Qd~iO#o~~", 25, 121), new LocalTokenGroup("d~RQYZXrs^~^O#y~~cO#z~~", 19, 132)],
+  tokenData: ":WlRzOX#uXY$^YZ%fZ]#u]^$^^p#upq$^qr%yrs&usu#uuv'Yvx#uxy(Wyz(kz{)O{|)||}*z}!O+_!O!P,s!P!Q-W!Q![/Y![!]3O!]!^3c!^!_3v!_!`4t!`!a6Y!a!b7W!b!c#u!c!}7k!}#O8h#O#P#u#P#Q8{#Q#R#u#R#S7k#S#T#u#T#o7k#o#p9`#p#q#u#q#r9s#r;'S#u;'S;=`$W<%lO#uW#zS#cWOY#uZ;'S#u;'S;=`$W<%lO#uW$ZP;=`<%l#ul$eY#cW#idOX#uXY$^YZ%TZ]#u]^$^^p#upq$^q;'S#u;'S;=`$W<%lO#ud%YS#idXY%TYZ%T]^%Tpq%Tl%mS#dW#idXY%TYZ%T]^%Tpq%T[&OU#cWOY#uZ!_#u!_!`&b!`;'S#u;'S;=`$W<%lO#u[&iS!RS#cWOY#uZ;'S#u;'S;=`$W<%lO#u[&|S#rS#cWOY#uZ;'S#u;'S;=`$W<%lO#u['aU!_S#cWOY#uZ!_#u!_!`'s!`;'S#u;'S;=`$W<%lO#u['zSyS#cWOY#uZ;'S#u;'S;=`$W<%lO#u[(_SWS#cWOY#uZ;'S#u;'S;=`$W<%lO#u[(rScS#cWOY#uZ;'S#u;'S;=`$W<%lO#u[)VU!]S#cWOY#uZ!_#u!_!`)i!`;'S#u;'S;=`$W<%lO#u[)pSwS#cWOY#uZ;'S#u;'S;=`$W<%lO#u[*TU!YS#cWOY#uZ!_#u!_!`*g!`;'S#u;'S;=`$W<%lO#u[*nSuS#cWOY#uZ;'S#u;'S;=`$W<%lO#u[+RSbS#cWOY#uZ;'S#u;'S;=`$W<%lO#ul+fV!ZS#cWOY#uZ!_#u!_!`+{!`!a,`!a;'S#u;'S;=`$W<%lO#u[,SSvS#cWOY#uZ;'S#u;'S;=`$W<%lO#uh,gSd`#cWOY#uZ;'S#u;'S;=`$W<%lO#u[,zS{S#cWOY#uZ;'S#u;'S;=`$W<%lO#ul-_Y!^S#cWOY#uZz#uz{-}{!P#u!P!Q.b!Q!_#u!_!`.u!`;'S#u;'S;=`$W<%lO#ul.US#ed#cWOY#uZ;'S#u;'S;=`$W<%lO#ul.iS#bd#cWOY#uZ;'S#u;'S;=`$W<%lO#u[.|SxS#cWOY#uZ;'S#u;'S;=`$W<%lO#u[/aW#cWoSOY#uZ!O#u!O!P/y!P!Q#u!Q![/Y![;'S#u;'S;=`$W<%lO#u[0OU#cWOY#uZ!Q#u!Q![0b![;'S#u;'S;=`$W<%lO#u[0iY#cWpSOY#uZ!Q#u!Q![0b![!g#u!g!h1X!h#X#u#X#Y1X#Y;'S#u;'S;=`$W<%lO#u[1^Y#cWOY#uZ{#u{|1||}#u}!O1|!O!Q#u!Q![2e![;'S#u;'S;=`$W<%lO#u[2RU#cWOY#uZ!Q#u!Q![2e![;'S#u;'S;=`$W<%lO#u[2lU#cWpSOY#uZ!Q#u!Q![2e![;'S#u;'S;=`$W<%lO#u[3VSXS#cWOY#uZ;'S#u;'S;=`$W<%lO#u[3jS!tS#cWOY#uZ;'S#u;'S;=`$W<%lO#u[3}U!VS#cWOY#uZ!_#u!_!`4a!`;'S#u;'S;=`$W<%lO#u[4hS!WS#cWOY#uZ;'S#u;'S;=`$W<%lO#u[4{VjS#cWOY#uZ!_#u!_!`5b!`!a5u!a;'S#u;'S;=`$W<%lO#u[5iS!QS#cWOY#uZ;'S#u;'S;=`$W<%lO#u[5|S!qS#cWOY#uZ;'S#u;'S;=`$W<%lO#u[6aU!TS#cWOY#uZ!_#u!_!`6s!`;'S#u;'S;=`$W<%lO#u[6zS!US#cWOY#uZ;'S#u;'S;=`$W<%lO#u[7_S}S#cWOY#uZ;'S#u;'S;=`$W<%lO#u[7r[US#cWOY#uZ!Q#u!Q![7k![!c#u!c!}7k!}#R#u#R#S7k#S#T#u#T#o7k#o;'S#u;'S;=`$W<%lO#u[8oS`S#cWOY#uZ;'S#u;'S;=`$W<%lO#u[9SSaS#cWOY#uZ;'S#u;'S;=`$W<%lO#ul9gSfd#cWOY#uZ;'S#u;'S;=`$W<%lO#u[9zS!mS#cWOY#uZ;'S#u;'S;=`$W<%lO#u",
+  tokenizers: [2, 3, 4, new LocalTokenGroup("j~RQYZXz{^~^O#g~~aP!P!Qd~iO#h~~", 25, 114), new LocalTokenGroup("d~RQYZXrs^~^O#t~~cO#u~~", 19, 127)],
   topRules: {"Program":[0,3]},
   specialized: [{term: 6, get: (value) => spec_IDENTIFIER[value] || -1}],
-  tokenPrec: 1544
+  tokenPrec: 1540
 })
