+++ Frontend/src/lib/codemirror/parser/parser.terms.js
@@ -3,111 +3,104 @@ export const
   SINGLELINE_COMMENT = 1,
   MULTILINE_COMMENT = 2,
   Program = 3,
-  Stmt = 4,
+  Top_level_stmt = 4,
   Func = 5,
   IDENTIFIER = 6,
   fn = 7,
   LPAREN = 8,
-  Param = 9,
-  COLON = 10,
-  Type_identifier = 11,
-  TYPE_LITERAL = 12,
-  int = 13,
-  float = 14,
-  bool = 15,
-  str = 16,
-  _void = 17,
-  LBRACKET = 18,
-  RBRACKET = 19,
-  COMMA = 20,
-  RPAREN = 21,
-  Return_type = 22,
-  ARROW = 23,
-  Block = 24,
-  LBRACE = 25,
-  Block_valid_stmt = 26,
-  Decl_stmt = 27,
-  _var = 28,
-  EQUAL = 29,
-  Expr = 30,
-  Assign_expr = 31,
-  Primary = 32,
-  INT_LITERAL = 33,
-  FLOAT_LITERAL = 34,
-  BOOL_LITERAL = 35,
-  _true = 36,
-  _false = 37,
-  STR_LITERAL = 38,
-  ASSIGN_OP = 39,
-  PLUS_EQUAL = 40,
-  MINUS_EQUAL = 41,
-  STAR_EQUAL = 42,
-  SLASH_EQUAL = 43,
-  PERCENT_EQUAL = 44,
-  Ternary_expr = 45,
-  QUESTION = 46,
-  Unary_expr = 47,
-  PREFIX_UNARY_OP = 48,
-  not = 49,
-  MINUS = 50,
-  Type_cast = 51,
-  as = 52,
-  Call_expr = 53,
-  Subscript_expr = 54,
-  Index_assign = 55,
-  Direct_member_access = 56,
-  DOT = 57,
-  Struct_initialization = 58,
-  Maybe_struct_initialization_list = 59,
-  Struct_initialization_list = 60,
-  RBRACE = 61,
-  Array_initialization = 62,
-  Match = 63,
-  match = 64,
-  Maybe_match_arms = 65,
-  Match_arms = 66,
-  FAT_ARROW = 67,
-  Match_else_arm = 68,
-  _else = 69,
-  Binary_expr = 70,
-  EQUALITY_OP = 71,
-  EQUAL_EQUAL = 72,
-  BANG_EQUAL = 73,
-  COMPARISON_OP = 74,
-  GREATER = 75,
-  GREATER_EQUAL = 76,
-  LESS = 77,
-  LESS_EQUAL = 78,
-  TERM_OP = 79,
-  PLUS = 80,
-  FACTOR_OP = 81,
-  STAR = 82,
-  SLASH = 83,
-  PERCENT = 84,
-  and = 85,
-  xor = 86,
-  or = 87,
-  Grouping = 88,
-  SEMICOLON = 89,
-  If_stmt = 90,
-  _if = 91,
-  Const_stmt = 92,
-  _const = 93,
-  Print_stmt = 94,
-  print = 95,
-  While_stmt = 96,
-  _while = 97,
-  For_stmt = 98,
-  _for = 99,
-  Return_stmt = 100,
-  _return = 101,
-  Break_stmt = 102,
-  _break = 103,
-  Continue_stmt = 104,
-  _continue = 105,
-  Expr_stmt = 106,
-  Type_stmt = 107,
-  type = 108,
-  Struct_decl = 109,
-  struct = 110,
-  Field_member = 111
+  COLON = 9,
+  Type_identifier = 10,
+  int = 11,
+  float = 12,
+  bool = 13,
+  str = 14,
+  _void = 15,
+  LBRACKET = 16,
+  RBRACKET = 17,
+  COMMA = 18,
+  RPAREN = 19,
+  ARROW = 20,
+  Block = 21,
+  LBRACE = 22,
+  Stmt = 23,
+  Decl_stmt = 24,
+  _var = 25,
+  EQUAL = 26,
+  Expr = 27,
+  Assign_expr = 28,
+  Primary = 29,
+  self = 30,
+  INT_LITERAL = 31,
+  FLOAT_LITERAL = 32,
+  _true = 33,
+  _false = 34,
+  STR_LITERAL = 35,
+  ASSIGN_OP = 36,
+  PLUS_EQUAL = 37,
+  MINUS_EQUAL = 38,
+  STAR_EQUAL = 39,
+  SLASH_EQUAL = 40,
+  PERCENT_EQUAL = 41,
+  Direct_member_access = 42,
+  DOT = 43,
+  Ternary_expr = 44,
+  QUESTION = 45,
+  Binary_expr = 46,
+  EQUALITY_OP = 47,
+  EQUAL_EQUAL = 48,
+  BANG_EQUAL = 49,
+  COMPARISON_OP = 50,
+  GREATER = 51,
+  GREATER_EQUAL = 52,
+  LESS = 53,
+  LESS_EQUAL = 54,
+  TERM_OP = 55,
+  PLUS = 56,
+  MINUS = 57,
+  FACTOR_OP = 58,
+  STAR = 59,
+  SLASH = 60,
+  PERCENT = 61,
+  and = 62,
+  xor = 63,
+  or = 64,
+  Unary_expr = 65,
+  PREFIX_UNARY_OP = 66,
+  not = 67,
+  Type_cast = 68,
+  as = 69,
+  Call_expr = 70,
+  Method_call = 71,
+  Subscript_expr = 72,
+  Index_assign = 73,
+  Struct_initialization = 74,
+  RBRACE = 75,
+  Array_initialization = 76,
+  Match = 77,
+  match = 78,
+  FAT_ARROW = 79,
+  _else = 80,
+  Grouping = 81,
+  SEMICOLON = 82,
+  If_stmt = 83,
+  _if = 84,
+  Const_stmt = 85,
+  _const = 86,
+  Print_stmt = 87,
+  print = 88,
+  While_stmt = 89,
+  _while = 90,
+  For_stmt = 91,
+  _for = 92,
+  Return_stmt = 93,
+  _return = 94,
+  Break_stmt = 95,
+  _break = 96,
+  Continue_stmt = 97,
+  _continue = 98,
+  Expr_stmt = 99,
+  Type_stmt = 100,
+  type = 101,
+  Struct_decl = 102,
+  struct = 103,
+  Prop_decl = 104
