+++ Frontend/src/routes/+page.svelte
@@ -5,7 +5,7 @@
 	import { getCode } from '$lib/getCode';
 	import { lex } from '$lib/lex';
 	import { textEditorCode } from '$lib/text-editor-code';
-	import { getExtensions } from '$lib/codemirror/codemirror-helpers'
+	import { get_extensions } from '$lib/codemirror/codemirror_helpers'
 	import { oneDarkTheme } from '@codemirror/theme-one-dark';
 	import { Label, Select } from 'flowbite-svelte';
 	import { onDestroy, onMount } from 'svelte';
@@ -139,7 +139,7 @@
 		<div>
 			<CodeMirror
 				extensions={[
-					...getExtensions(), // Apply extensions for additional functionality like syntax highlighting.
+					...get_extensions(), // Apply extensions for additional functionality like syntax highlighting.
 					oneDarkTheme,
 				]}
 				bind:value={$textEditorCode}
