+++ Frontend/src/routes/playground/+page.svelte
@@ -3,7 +3,7 @@
 	import { consoleOutput } from '$lib/console-output';
 	import { currentCompiledWat } from '$lib/current-compiled-wat';
 	import { textEditorCode } from '$lib/text-editor-code';
-	import { getExtensions } from '$lib/codemirror/codemirror-helpers'
+	import { get_extensions } from '$lib/codemirror/codemirror_helpers'
 	import { oneDarkTheme } from '@codemirror/theme-one-dark';
 	import { TabItem, Tabs } from 'flowbite-svelte';
 	import CodeMirror from 'svelte-codemirror-editor';
@@ -28,7 +28,7 @@
 		<div class="grid w-full grid-cols-3">
 			<CodeMirror
 				extensions={[
-					...getExtensions(), // Apply extensions for additional functionality like syntax highlighting.
+					...get_extensions(), // Apply extensions for additional functionality like syntax highlighting.
 					oneDarkTheme,
 				]}
 				basic
